+++ skfem/assembly/form/form.py
@@ -1,6 +1,7 @@
 from typing import Any, Callable, Optional, Tuple, Union
 
 from numpy import ndarray
+from scipy.sparse import coo_matrix
 
 from .form_parameters import FormParameters
 from ..basis import Basis
@@ -21,9 +22,20 @@ class Form:
         raise NotImplementedError
 
     @staticmethod
-    def parameters(w: Optional[Any], ubasis: Basis) -> FormParameters:
+    def parameters(w: Optional[Any], u: Basis) -> FormParameters:
         if type(w) is list:
             w = zip(*w)
         elif type(w) is ndarray:
             w = (w, None, None)
-        return FormParameters(*w, **ubasis.default_parameters())
+        return FormParameters(*w, **u.default_parameters())
+
+    @staticmethod
+    def _assemble_scipy_matrix(data, rows, cols, shape=None):
+        K = coo_matrix((data, (rows, cols)), shape=shape)
+        K.eliminate_zeros()
+        return K.tocsr()
+
+    @staticmethod
+    def _assemble_numpy_vector(data, rows, cols, shape=None):
+        return coo_matrix((data, (rows, cols)),
+                          shape=shape).toarray().T[0]